requirements.txt
# python3.12 -m pip install -r requirements.txt
mysql-connector-python==8.0.29
redis==5.0.8